+++ src/main/java/org/opentripplanner/gtfs/GenerateTripPatternsOperation.java
@@ -7,17 +7,9 @@ import org.opentripplanner.graph_builder.DataImportIssueStore;
 import org.opentripplanner.graph_builder.issues.GTFSModeNotSupported;
 import org.opentripplanner.graph_builder.issues.TripDegenerate;
 import org.opentripplanner.graph_builder.issues.TripUndefinedService;
-import org.opentripplanner.model.Direction;
-import org.opentripplanner.model.FeedScopedId;
-import org.opentripplanner.model.Frequency;
-import org.opentripplanner.model.Route;
-import org.opentripplanner.model.StopPattern;
-import org.opentripplanner.model.StopTime;
-import org.opentripplanner.model.Trip;
-import org.opentripplanner.model.TripPattern;
+import org.opentripplanner.model.*;
 import org.opentripplanner.model.impl.OtpTransitServiceBuilder;
 import org.opentripplanner.routing.trippattern.Deduplicator;
-import org.opentripplanner.routing.trippattern.FrequencyEntry;
 import org.opentripplanner.routing.trippattern.TripTimes;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -27,6 +19,7 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
+import java.util.stream.Collectors;
 
 /**
  * This class is responsible for generating trip patterns when loading GTFS data.
@@ -42,7 +35,7 @@ public class GenerateTripPatternsOperation {
     private final Set<FeedScopedId> calendarServiceIds;
 
     private final Multimap<StopPattern, TripPattern> tripPatterns;
-    private final ListMultimap<Trip, Frequency> frequenciesForTrip = ArrayListMultimap.create();
+    private final Map<Trip, List<Frequency>> frequenciesForTrip;
 
     private int tripCount = 0;
     private int freqCount = 0;
@@ -57,10 +50,12 @@ public class GenerateTripPatternsOperation {
         this.deduplicator = deduplicator;
         this.calendarServiceIds = calendarServiceIds;
         this.tripPatterns = transitDaoBuilder.getTripPatterns();
+        this.frequenciesForTrip = transitDaoBuilder.getFrequencies()
+                .stream()
+                .collect(Collectors.groupingBy(Frequency::getTrip));
     }
 
     public void run() {
-        collectFrequencyByTrip();
 
         final Collection<Trip> trips = transitDaoBuilder.getTripsById().values();
         final int tripsSize = trips.size();
@@ -86,17 +81,6 @@ public class GenerateTripPatternsOperation {
         return scheduledCount > 0;
     }
 
-    /**
-     * First, record which trips are used by one or more frequency entries.
-     * These trips will be ignored for the purposes of non-frequency routing, and
-     * all the frequency entries referencing the same trip can be added at once to the same
-     * Timetable/TripPattern.
-     */
-    private void collectFrequencyByTrip() {
-        for(Frequency freq : transitDaoBuilder.getFrequencies()) {
-            frequenciesForTrip.put(freq.getTrip(), freq);
-        }
-    }
 
     private void buildTripPatternForTrip(Trip trip) {
         // TODO: move to a validator module
@@ -128,25 +112,63 @@ public class GenerateTripPatternsOperation {
             stopPattern, trip.getRoute(), direction
         );
 
-        // Create a TripTimes object for this list of stoptimes, which form one trip.
-        TripTimes tripTimes = new TripTimes(trip, stopTimes, deduplicator);
-
         // If this trip is referenced by one or more lines in frequencies.txt, wrap it in a FrequencyEntry.
         List<Frequency> frequencies = frequenciesForTrip.get(trip);
         if (frequencies != null && !(frequencies.isEmpty())) {
-            for (Frequency freq : frequencies) {
-                tripPattern.add(new FrequencyEntry(freq, tripTimes));
+            // Create a TripTimes object for this list of stoptimes, which form one trip.
+            var tripTimesMap = new HashMap<Integer, TripTimes>();
+                for (Frequency freq : frequencies) {
+                    TripTimes tripTimes = tripTimesMap.computeIfAbsent(
+                            freq.getExactTimes() ==1 ? 0: freq.getHeadwaySecs(),
+                            headway -> new TripTimes(trip, shiftStopTimes(stopTimes, headway), deduplicator)
+                    );
+                    int timeShift = timeShiftForFrequencyBasedTrip(tripTimes, freq);
+                    while (timeShift <= freq.getEndTime()) {
+                        tripPattern.add(new TripTimes(tripTimes, freq, timeShift));
+                        timeShift += freq.getHeadwaySecs();
+                    }
                 freqCount++;
+            }
+            if (tripPattern.getTrips().isEmpty()) {
+                LOG.error("Frequency expansion for trip {} via route {} for service {} yielded 0 patterns",
+                        trip.getId(), trip.getRoute().getId(), trip.getServiceId());
             }
             // TODO replace: createGeometry(graph, trip, stopTimes, hops);
         }
 
         // This trip was not frequency-based. Add the TripTimes directly to the TripPattern's scheduled timetable.
         else {
+            TripTimes tripTimes = new TripTimes(trip, stopTimes, deduplicator);
             tripPattern.add(tripTimes);
             scheduledCount++;
         }
     }
+
+    /**
+     else {
+     * According to the GTFS specification, the first trip depart at the frequency
+     tripPattern.add(tripTimes);
+     * start-time and the trip-pattern time-shift is the time of the first stop
+     scheduledCount++;
+     * arrival. So, we compute the delta between the arrival and departure and
+     }
+     * subtract that from the frequency start-time to get the initial time-shift value.
+     */
+    private int timeShiftForFrequencyBasedTrip(TripTimes tripTimes, Frequency frequency) {
+        int delta = tripTimes.getDepartureTime(0) - tripTimes.getArrivalTime(0);
+        return frequency.getStartTime() - delta;
+    }
+
+    private static Collection<StopTime> shiftStopTimes(Collection<StopTime> stopTimes, int timeShiftArrival) {
+        return timeShiftArrival == 0 ? stopTimes : stopTimes.stream()
+                .map(it -> {
+                    var st = new StopTime(it);
+                    st.setArrivalTime(it.getArrivalTime() + timeShiftArrival);
+                    return st;
+                })
+                .collect(Collectors.toList());
+    }
+
 
     private TripPattern findOrCreateTripPattern(StopPattern stopPattern, Route route, Direction direction) {
         for(TripPattern tripPattern : tripPatterns.get(stopPattern)) {